// web/src/lib/api/generated/models/internalApiHandlersEnrichmentQueueItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface InternalApiHandlersEnrichmentQueueItem {
  attempts?: number
  created_at?: string
  error_category?: string
  error_message?: string
  id?: number
  library_id?: number
  max_attempts?: number
  media_id?: number
  media_type?: string
  priority?: number
  stage?: string
  status?: string
  title?: string
  updated_at?: string
}